create newCommentForm component and new logic
