+++ components/PairingItem.js
@@ -1,6 +1,5 @@
 import { useState, useEffect } from "react";
 import styled from "styled-components";
-import { nanoid } from "nanoid";
 import {
   Card,
   Ingredients,
@@ -14,13 +13,9 @@ import {
 } from "@/_styles";
 import { flavorColors } from "@/utils";
 import ingredientsData from "@/assets/ingredients.json";
-import CommentPopup from "./CommentPopup";
 
-const PairingItem = ({ pairing }) => {
+const PairingItem = ({ pairing, setShow }) => {
   const [favorited, setFavorited] = useState(false);
-  const [showCommentPopup, setShowCommentPopup] = useState(false);
-  const [comments, setComments] = useState([]);
-  const [editingComment, setEditingComment] = useState(null);
 
   const [ingredients, setIngredients] = useState([]);
 
@@ -33,32 +28,6 @@ const PairingItem = ({ pairing }) => {
 
   const toggleFavorite = () => {
     setFavorited(!favorited);
-  };
-
-  const handleCommentSubmit = (comment, commentId) => {
-    if (commentId) {
-      setComments(
-        comments.map((comment) =>
-          comment.id === commentId ? { ...comment, text: comment } : c
-        )
-      );
-    } else {
-      const newComment = { id: nanoid(), text: comment };
-      setComments([...comments, newComment]);
-    }
-    setShowCommentPopup(false);
-    setEditingComment(null);
-  };
-
-  const handleEdit = (commentId) => {
-    const commentToEdit = comments.find((comment) => comment.id === commentId);
-    setEditingComment(commentToEdit);
-    setShowCommentPopup(true);
-  };
-
-  const handleDelete = (commentId) => {
-    setComments(comments.filter((comment) => comment.id !== commentId));
-    setShowCommentPopup(false);
   };
 
   return (
@@ -92,28 +61,8 @@ const PairingItem = ({ pairing }) => {
             #{ingredient.flavor}
           </Flavors>
         ))}
-        <CommentEmoji onClick={() => setShowCommentPopup(true)}>
-          💬
-        </CommentEmoji>
+        <CommentEmoji onClick={() => setShow(true)}>💬</CommentEmoji>
       </CardFooter>
-      <CommentsSection>
-        {comments.map((comment) => (
-          <Comment key={comment.id}>
-            {comment.text}{" "}
-            <EditButton onClick={() => handleEdit(comment.id)}>Edit</EditButton>
-          </Comment>
-        ))}
-      </CommentsSection>
-      <CommentPopup
-        show={showCommentPopup}
-        onClose={() => {
-          setShowCommentPopup(false);
-          setEditingComment(null);
-        }}
-        onSubmit={handleCommentSubmit}
-        commentToEdit={editingComment}
-        onDelete={handleDelete}
-      />
     </Card>
   );
 };
@@ -123,42 +72,4 @@ export default PairingItem;
 const CommentEmoji = styled.span`
   cursor: pointer;
   margin-left: auto;
-`;
-
-const CommentsSection = styled.div`
-  margin-top: 20px;
-  padding: 10px;
-  background: #f9f9f9;
-  border-radius: 8px;
-`;
-
-const Comment = styled.div`
-  background: #ffffff;
-  padding: 12px;
-  border-radius: 8px;
-  margin-top: 10px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-`;
-
-const EditButton = styled.button`
-  background: none;
-  border: none;
-  color: #007bff;
-  cursor: pointer;
-  margin-left: 8px;
-  font-size: 14px;
-  padding: 4px 8px;
-  border-radius: 4px;
-  transition: background-color 0.3s, color 0.3s;
-  &:hover {
-    background-color: #007bff;
-    color: #fff;
-  }
-  &:focus {
-    outline: none;
-    box-shadow: 0 0 0 2px rgba(0, 123, 255, 0.5);
-  }
-  &:active {
-    background-color: #0056b3;
-  }
 `;
